Fixed warning.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@32987 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/installer/UnzipEngine.cpp
@@ -184,7 +184,7 @@ UnzipEngine::UnzipPackage()
 		ssize_t read = descriptionFile.Read(buffer, bufferSize);
 		if (read > 0) {
 			int32 length = copyrightAttr.Length();
-			if (read < bufferSize)
+			if (read < (ssize_t)bufferSize)
 				buffer[read] = '\0';
 			int32 bufferLength = strlen(buffer);
 				// Should be "read", but maybe we have a zero in the
